Revert "Add GUI app with drag-and-drop and MIT license"
