--- .gitignore
@@ -1,5 +1,0 @@
-__pycache__/
-*.pyc
-build/
-dist/
-*.spec
